Update 03_PopUpFloat_Translater.ipynb
TempConversion method already had logic for an alternative cal bath equation (log10 vs log) but it had not been implemented in the function calls throughout the code.  add this to routines and set to version=2 which is the more recent version.  reprocessing of old units will need to set the version appropriatly